[Gardening] Split Bitcode from Bitstream from BitstreamReader
Make a logical distinction between Bitcode the format, Bitstream the container, and all of the data therein by splitting the bitstream reader infrastructure into its own file.
