+++ Sources/TSCUtility/CMakeLists.txt
@@ -12,6 +12,7 @@ add_library(TSCUtility
   ArgumentParserShellCompletion.swift
   Bits.swift
   Bitstream.swift
+  BitstreamReader.swift
   BuildFlags.swift
   CollectionExtensions.swift
   Diagnostics.swift
